crypto/pbkdf2_sha1: add a PBKDF2-SHA-1 module
The existing crypto/pbkdf2 uses SHA-512, which is fine, but will not
work with the SCRAM-SHA-1 authentication mechanism.

In addition, a better hash format is used, allowing to clearly
distinguish between pbkdf2 and pbkdf2_sha1 hashes, and avoiding the need
to hardcode SALTLEN and ROUNDS in the future SCRAM-SHA-1 module.